requirements/recommended.txt
-r required.txt

# For better performance
# 0.11.0 is broken using pip, later versions are fixed
python-Levenshtein>=0.10.2,!=0.11.0

# For management commands
Flask-Script>=0.5.0
